Add files via upload
Robot Pointing on a Wall 

Circle with gradient
Gradient that fades to transparency
Comillas Negras
Task-03

03

Local Store E-commerce Platform

Create an e-commerce website for a local
store in your area, enabling customers to
browse and purchase products online. 

A product that stands out from the platform
Listings with photos, descriptions and prices,
And the shopping cart function.